+++ lib/appbar.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(const MaterialApp(
+    title: 'AppBar',
+    home: BaseAppBar(),
+  ));
+}
+
+class BaseAppBar extends StatelessWidget implements PreferredSizeWidget {
+  const BaseAppBar({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      title: const Text('StitchVault'),
+      actions: <Widget>[
+        IconButton(
+          icon: const Icon(Icons.home),
+          tooltip: 'Home Page',
+          onPressed: () {
+            ScaffoldMessenger.of(context).showSnackBar(
+                const SnackBar(content: Text('This is a snackbar')));
+          },
+        )
+      ],
+      backgroundColor: Colors.green,
+    );
+  }
+
+  @override
+  Size get preferredSize => Size.fromHeight(preferredSize.height);
+}
