+++ lib/homepage.dart
@@ -1,371 +1,102 @@
 import 'package:flutter/material.dart';
+import 'appbar.dart';
 
-/// Flutter code sample for [NavigationBar] with nested [Navigator] destinations.
+/// Flutter code sample for [BottomNavigationBar].
 
-void main() {
-  runApp(const MaterialApp(home: Home()));
-}
+void main() => runApp(const Home());
 
-class Home extends StatefulWidget {
+class Home extends StatelessWidget {
   const Home({super.key});
 
   @override
-  State<Home> createState() => _HomeState();
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      home: BottomNavigationBarExample(),
+    );
+  }
 }
 
-class _HomeState extends State<Home> with TickerProviderStateMixin<Home> {
-  static const List<Destination> allDestinations = <Destination>[
-    Destination(0, 'Users', Icons.person, Colors.blue),
-    Destination(1, 'Designs', Icons.palette, Colors.orange),
-    Destination(2, 'Work Entries', Icons.business, Colors.red),
-    Destination(3, 'Payments', Icons.currency_rupee_outlined, Colors.green),
+class BottomNavigationBarExample extends StatefulWidget {
+  const BottomNavigationBarExample({super.key});
+
+  @override
+  State<BottomNavigationBarExample> createState() =>
+      _BottomNavigationBarExampleState();
+}
+
+class _BottomNavigationBarExampleState
+    extends State<BottomNavigationBarExample> {
+  int _selectedIndex = 0;
+  static const TextStyle optionStyle =
+      TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
+  static const List<Widget> _widgetOptions = <Widget>[
+    Text(
+      'Index 0: Users',
+      style: optionStyle,
+    ),
+    Text(
+      'Index 1: Designs',
+      style: optionStyle,
+    ),
+    Text(
+      'Index 2: Work Entries',
+      style: optionStyle,
+    ),
+    Text(
+      'Index 3: Payment',
+      style: optionStyle,
+    ),
   ];
 
-  late final List<GlobalKey<NavigatorState>> navigatorKeys;
-  late final List<GlobalKey> destinationKeys;
-  late final List<AnimationController> destinationFaders;
-  late final List<Widget> destinationViews;
-  int selectedIndex = 0;
-
-  AnimationController buildFaderController() {
-    final AnimationController controller = AnimationController(
-        vsync: this, duration: const Duration(milliseconds: 200));
-    controller.addStatusListener((AnimationStatus status) {
-      if (status == AnimationStatus.dismissed) {
-        setState(() {}); // Rebuild unselected destinations offstage.
-      }
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
     });
-    return controller;
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    navigatorKeys = List<GlobalKey<NavigatorState>>.generate(
-        allDestinations.length, (int index) => GlobalKey()).toList();
-    destinationFaders = List<AnimationController>.generate(
-        allDestinations.length, (int index) => buildFaderController()).toList();
-    destinationFaders[selectedIndex].value = 1.0;
-    destinationViews = allDestinations.map((Destination destination) {
-      return FadeTransition(
-        opacity: destinationFaders[destination.index]
-            .drive(CurveTween(curve: Curves.fastOutSlowIn)),
-        child: DestinationView(
-          destination: destination,
-          navigatorKey: navigatorKeys[destination.index],
-        ),
-      );
-    }).toList();
-  }
-
-  @override
-  void dispose() {
-    for (final AnimationController controller in destinationFaders) {
-      controller.dispose();
-    }
-    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return WillPopScope(
-      onWillPop: () async {
-        final NavigatorState navigator =
-            navigatorKeys[selectedIndex].currentState!;
-        if (!navigator.canPop()) {
-          return true;
-        }
-        navigator.pop();
-        return false;
-      },
-      child: Scaffold(
-        body: SafeArea(
-          top: false,
-          child: Stack(
-            fit: StackFit.expand,
-            children: allDestinations.map((Destination destination) {
-              final int index = destination.index;
-              final Widget view = destinationViews[index];
-              if (index == selectedIndex) {
-                destinationFaders[index].forward();
-                return Offstage(offstage: false, child: view);
-              } else {
-                destinationFaders[index].reverse();
-                if (destinationFaders[index].isAnimating) {
-                  return IgnorePointer(child: view);
-                }
-                return Offstage(child: view);
-              }
-            }).toList(),
-          ),
-        ),
-        bottomNavigationBar: NavigationBar(
-          selectedIndex: selectedIndex,
-          onDestinationSelected: (int index) {
-            setState(() {
-              selectedIndex = index;
-            });
-          },
-          destinations: allDestinations.map((Destination destination) {
-            return NavigationDestination(
-              icon: Icon(destination.icon, color: destination.color),
-              label: destination.title,
-            );
-          }).toList(),
-        ),
-      ),
-    );
-  }
-}
-
-class Destination {
-  const Destination(this.index, this.title, this.icon, this.color);
-  final int index;
-  final String title;
-  final IconData icon;
-  final MaterialColor color;
-}
-
-class RootPage extends StatelessWidget {
-  const RootPage({super.key, required this.destination});
-
-  final Destination destination;
-
-  Widget _buildDialog(BuildContext context) {
-    return AlertDialog(
-      title: Text('${destination.title} AlertDialog'),
-      actions: <Widget>[
-        TextButton(
-          onPressed: () {
-            Navigator.pop(context);
-          },
-          child: const Text('OK'),
-        ),
-      ],
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final TextStyle headlineSmall = Theme.of(context).textTheme.headlineSmall!;
-    final ButtonStyle buttonStyle = ElevatedButton.styleFrom(
-      backgroundColor: destination.color,
-      visualDensity: VisualDensity.comfortable,
-      padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 16),
-      textStyle: headlineSmall,
-    );
-
     return Scaffold(
       appBar: AppBar(
-        title: Text('StitchVault'),
-        backgroundColor: destination.color,
+        title: const Text('StitchVault'),
+        actions: <Widget>[
+          IconButton(
+            icon: const Icon(Icons.home),
+            tooltip: 'Home Page',
+            onPressed: () {
+              ScaffoldMessenger.of(context).showSnackBar(
+                  const SnackBar(content: Text('This is a snackbar')));
+            },
+          )
+        ],
       ),
-      backgroundColor: destination.color[50],
       body: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            ElevatedButton(
-              style: buttonStyle,
-              onPressed: () {
-                Navigator.pushNamed(context, '/list');
-              },
-              child: const Text('Push /list'),
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              style: buttonStyle,
-              onPressed: () {
-                showDialog(
-                  context: context,
-                  useRootNavigator: false,
-                  builder: _buildDialog,
-                );
-              },
-              child: const Text('Local Dialog'),
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              style: buttonStyle,
-              onPressed: () {
-                showDialog(
-                  context: context,
-                  useRootNavigator: true,
-                  builder: _buildDialog,
-                );
-              },
-              child: const Text('Root Dialog'),
-            ),
-            const SizedBox(height: 16),
-            Builder(
-              builder: (BuildContext context) {
-                return ElevatedButton(
-                  style: buttonStyle,
-                  onPressed: () {
-                    showBottomSheet(
-                      context: context,
-                      builder: (BuildContext context) {
-                        return Container(
-                          padding: const EdgeInsets.all(16),
-                          width: double.infinity,
-                          child: Text(
-                            '${destination.title} BottomSheet\n'
-                            'Tap the back button to dismiss',
-                            style: headlineSmall,
-                            softWrap: true,
-                            textAlign: TextAlign.center,
-                          ),
-                        );
-                      },
-                    );
-                  },
-                  child: const Text('Local BottomSheet'),
-                );
-              },
-            ),
-          ],
-        ),
+        child: _widgetOptions.elementAt(_selectedIndex),
       ),
-    );
-  }
-}
-
-class ListPage extends StatelessWidget {
-  const ListPage({super.key, required this.destination});
-
-  final Destination destination;
-
-  @override
-  Widget build(BuildContext context) {
-    const int itemCount = 50;
-    final ButtonStyle buttonStyle = OutlinedButton.styleFrom(
-      foregroundColor: destination.color,
-      fixedSize: const Size.fromHeight(128),
-      textStyle: Theme.of(context).textTheme.headlineSmall,
-    );
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('${destination.title} ListPage - /list'),
-        backgroundColor: destination.color,
-      ),
-      backgroundColor: destination.color[50],
-      body: SizedBox.expand(
-        child: ListView.builder(
-          itemCount: itemCount,
-          itemBuilder: (BuildContext context, int index) {
-            return Padding(
-              padding: const EdgeInsets.symmetric(vertical: 4, horizontal: 8),
-              child: OutlinedButton(
-                style: buttonStyle.copyWith(
-                  backgroundColor: MaterialStatePropertyAll<Color>(
-                    Color.lerp(destination.color[100], Colors.white,
-                        index / itemCount)!,
-                  ),
-                ),
-                onPressed: () {
-                  Navigator.pushNamed(context, '/text');
-                },
-                child: Text('Push /text [$index]'),
-              ),
-            );
-          },
-        ),
-      ),
-    );
-  }
-}
-
-class TextPage extends StatefulWidget {
-  const TextPage({super.key, required this.destination});
-
-  final Destination destination;
-
-  @override
-  State<TextPage> createState() => _TextPageState();
-}
-
-class _TextPageState extends State<TextPage> {
-  late final TextEditingController textController;
-
-  @override
-  void initState() {
-    super.initState();
-    textController = TextEditingController(text: 'Sample Text');
-  }
-
-  @override
-  void dispose() {
-    textController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('${widget.destination.title} TextPage - /list/text'),
-        backgroundColor: widget.destination.color,
-      ),
-      backgroundColor: widget.destination.color[50],
-      body: Container(
-        padding: const EdgeInsets.all(32.0),
-        alignment: Alignment.center,
-        child: TextField(
-          controller: textController,
-          style: theme.primaryTextTheme.headlineMedium?.copyWith(
-            color: widget.destination.color,
+      bottomNavigationBar: BottomNavigationBar(
+        type: BottomNavigationBarType.fixed,
+        items: const <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Icon(Icons.person),
+            label: 'Users',
           ),
-          decoration: InputDecoration(
-            focusedBorder: UnderlineInputBorder(
-              borderSide: BorderSide(
-                color: widget.destination.color,
-                width: 3.0,
-              ),
-            ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.palette),
+            label: 'Designs',
           ),
-        ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.business),
+            label: 'Work Entries',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.currency_rupee_outlined),
+            label: 'Payment',
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        selectedItemColor: Colors.blue[800],
+        onTap: _onItemTapped,
       ),
-    );
-  }
-}
-
-class DestinationView extends StatefulWidget {
-  const DestinationView({
-    super.key,
-    required this.destination,
-    required this.navigatorKey,
-  });
-
-  final Destination destination;
-  final Key navigatorKey;
-
-  @override
-  State<DestinationView> createState() => _DestinationViewState();
-}
-
-class _DestinationViewState extends State<DestinationView> {
-  @override
-  Widget build(BuildContext context) {
-    return Navigator(
-      key: widget.navigatorKey,
-      onGenerateRoute: (RouteSettings settings) {
-        return MaterialPageRoute<void>(
-          settings: settings,
-          builder: (BuildContext context) {
-            switch (settings.name) {
-              case '/':
-                return RootPage(destination: widget.destination);
-              case '/list':
-                return ListPage(destination: widget.destination);
-              case '/text':
-                return TextPage(destination: widget.destination);
-            }
-            assert(false);
-            return const SizedBox();
-          },
-        );
-      },
     );
   }
 }
